Dashboard updates
Added a table of active cards, from their respective tables on the DB. Which can be found in the Dashboard page. In the dashboard page an admin can delete or edit the active cards. (Have to work a bit more on the edit function). Also turned the validation's (for both signup and login) and dashboard's php code in OOP. Working on a 'favorite cards feature' for users.